+++ src/plugins/chatboxviews/index.js
@@ -22,10 +22,7 @@ converse.plugins.add('converse-chatboxviews', {
         // ====================================
         // Refer to docs/source/configuration.rst for explanations of these
         // configuration settings.
-        api.settings.extend({
-            'animate': true,
-            'theme': 'default'
-        });
+        api.settings.extend({ 'animate': true });
 
         _converse.chatboxviews = new ChatBoxViews();