+++ webpack/webpack.headless.js
@@ -6,7 +6,7 @@ const { merge } = require("webpack-merge");
 module.exports = merge(common, {
     entry: "@converse/headless/headless.js",
     output: {
-        path: path.resolve(__dirname, 'src/headless/dist'), // Output path for generated bundles
+        path: path.resolve(__dirname, '../src/headless/dist'), // Output path for generated bundles
         filename: 'converse-headless.min.js',
         chunkFilename: '[name].js'
     },