+++ webpack/webpack.nodeps.js
@@ -48,7 +48,7 @@ module.exports = merge(common, {
                     loader: 'sass-loader',
                     options: {
                         sassOptions: {
-                            includePaths: [path.resolve(__dirname, 'node_modules/')]
+                            includePaths: [path.resolve(__dirname, '../node_modules/')]
                         }
                     }
                 }