+++ CHANGES.md
@@ -3,7 +3,13 @@
 ## 9.0.1 (2021-??-??)
 
 - Updated translations: lt
+- #1761: Add a new dark theme based on the [Dracula](https://draculatheme.com/) theme
 - #2751: Media not rendered when Converse runs in a browser extension
+- #2789: Implement new hook parseMessageForCommands for plugins to add custom
+
+### New hooks and events
+
+- New hook: [parseMessageForCommands](https://conversejs.org/docs/html/api/-_converse.html#event:parseMessageForCommands)
 
 ## 9.0.0 (2021-11-26)